fix(store): move test helper out of prod, harden repo error handling
- Remove `testing` import + OpenMem from store.go (production code);
  move OpenMem into internal/store/helpers_test.go (package store,
  test-only) so the production binary no longer pulls in the testing pkg
- Document UpsertTarget ON CONFLICT invariant: only appdata_paths is
  refreshed on conflict; id, created_at, and include_in_schedule are
  preserved — include_in_schedule is owned exclusively by SetInclude
- Replace `err == sql.ErrNoRows` with errors.Is in runs.go + settings.go
- Check RowsAffected in FinishRun and SetInclude; return descriptive
  errors when 0 rows affected (unknown run/container)
- Add scoped gosec G304/G306 exclusion in .golangci.yml for future
  internal/template and internal/paths file-I/O code (Wave C pre-empt)
