+++ internal/store/helpers_test.go
@@ -1,0 +1,18 @@
+package store
+
+import (
+	"database/sql"
+	"testing"
+)
+
+// OpenMem opens an in-memory SQLite database suitable for tests.
+// The database is closed automatically when t finishes.
+func OpenMem(t *testing.T) *sql.DB {
+	t.Helper()
+	db, err := Open(":memory:")
+	if err != nil {
+		t.Fatalf("store.OpenMem: %v", err)
+	}
+	t.Cleanup(func() { db.Close() })
+	return db
+}
